LU-15366 nrs: increase maximum rate for tbf rule
The maximum rpc rate for a tbf rule is 65535. This value could be
problematic for cluster with a large number of clients.

This patch uniformizes the usage of __u64 to store a rpc rate.
And changes the maximum rate for a tbf rule to 1000000 (1 rpc/us)

Reviewed-on: https://review.whamcloud.com/45838
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/include/lustre_nrs_tbf.h
@@ -102,7 +102,7 @@ struct nrs_tbf_client {
 	/** Generation of the rule matched. */
 	__u64				 tc_rule_generation;
 	/** Limit of RPC rate. */
-	u32				 tc_rpc_rate;
+	__u64				 tc_rpc_rate;
 	/** Time to wait for next token. */
 	__u64				 tc_nsecs;
 	/** RPC token number. */
@@ -169,7 +169,7 @@ struct nrs_tbf_rule {
 	/** Generic condition string of the rule. */
 	char				*tr_conds_str;
 	/** RPC/s limit. */
-	u32				 tr_rpc_rate;
+	__u64				 tr_rpc_rate;
 	/** Time to wait for next token. */
 	u64				 tr_nsecs_per_rpc;
 	/** Token bucket depth. */
